Refactor config.ts
